requirements-test.txt
pytest-celery==0.0.0
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-env==1.1.3
pytest==7.4.4
pre-commit==3.6.0
pytest-dotenv==0.5.2
black==23.12.1
fakeredis==2.20.1